scripts/report_generators/requirements.txt
numpy>=1.22.0
matplotlib>=3.5.0
reportlab>=3.6.0
pandas>=1.3.0
seaborn>=0.11.0
pillow>=9.0.0
sqlalchemy>=2.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
asyncio>=3.4.3
